map 64bit int anf bfloat16

+++ modules/objdetect/misc/python/test/test_objdetect_aruco.py
@@ -74,7 +74,8 @@ def projectChessboard(squaresX, squaresY, squareSize, imageSize, cameraMatrix, r
                                       squareCorners[0] + [0, squareSize, 0]))
 
             projectedCorners, _ = cv.projectPoints(squareCorners, rvec, tvec, cameraMatrix, distCoeffs)
-            projectedCorners = projectedCorners.astype(np.int64)
+            # fillPoly() requires CV_32S point coordinates
+            projectedCorners = projectedCorners.astype(np.int32)
             projectedCorners = projectedCorners.reshape(1, 4, 2)
             img = cv.fillPoly(img, [projectedCorners], 0)
